split out paginator
