+++ packages/datatrak-web/src/features/Survey/Components/SurveyScreenHeader.tsx
@@ -1,0 +1,43 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+import React from 'react';
+import styled from 'styled-components';
+import { Typography } from '@material-ui/core';
+
+const Header = styled.div`
+  padding: 1rem;
+  width: 100%;
+  border-bottom: 1px solid ${({ theme }) => theme.palette.divider};
+  ${({ theme }) => theme.breakpoints.up('sm')} {
+    padding: 1rem 2rem;
+  }
+  ${({ theme }) => theme.breakpoints.up('md')} {
+    padding: 1.375rem 2.75rem;
+  }
+`;
+
+const PageHeading = styled(Typography).attrs({
+  variant: 'h2',
+})`
+  ${({ theme }) => theme.breakpoints.down('sm')} {
+    font-size: 1rem;
+  }
+`;
+
+const PageDescription = styled(Typography)`
+  ${({ theme }) => theme.breakpoints.down('sm')} {
+    font-size: 0.875rem;
+  }
+`;
+
+export const SurveyScreenHeader = ({ heading, description }) => {
+  return (
+    <Header>
+      <PageHeading>{heading}</PageHeading>
+      <PageDescription>{description}</PageDescription>
+    </Header>
+  );
+};
