+++ packages/datatrak-web/src/features/Survey/Components/index.ts
@@ -8,3 +8,5 @@ export { SurveyQuestionGroup } from './SurveyQuestionGroup';
 export { SurveyToolbar } from './SurveyToolbar';
 export * from './SurveySideMenu';
 export { SurveyReviewSection } from './SurveyReviewSection';
+export { SurveyScreenHeader } from './SurveyScreenHeader';
+export { SurveyPaginator } from './SurveyPaginator';
