+++ packages/datatrak-web/src/features/Survey/SurveyContext/SurveyContext.tsx
@@ -138,6 +138,7 @@ export const useSurveyForm = () => {
     isSuccessScreen,
     isReviewScreen,
     isResponseScreen,
+    numberOfScreens,
     toggleSideMenu,
     setFormData,
     resetForm,
