+++ packages/datatrak-web/src/features/Survey/SurveyLayout.tsx
@@ -7,16 +7,13 @@ import React from 'react';
 import { Outlet, generatePath, useNavigate, useParams } from 'react-router';
 import { useFormContext } from 'react-hook-form';
 import styled from 'styled-components';
-import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos';
 import { Paper as MuiPaper } from '@material-ui/core';
 import { SpinningLoader } from '@tupaia/ui-components';
 import { SurveyParams } from '../../types';
 import { useSurveyForm } from './SurveyContext';
-import { SIDE_MENU_WIDTH, SurveySideMenu } from './Components';
-import { ROUTES } from '../../constants';
-import { Button } from '../../components';
+import { SIDE_MENU_WIDTH, SurveySideMenu, SurveyToolbar, SurveyPaginator } from './Components';
+import { HEADER_HEIGHT, ROUTES, SURVEY_TOOLBAR_HEIGHT } from '../../constants';
 import { useSubmitSurvey } from '../../api/mutations';
-import { useIsMobile } from '../../utils';
 import { getErrorsByScreen } from './utils';
 
 const ScrollableLayout = styled.div<{
@@ -31,6 +28,21 @@ const ScrollableLayout = styled.div<{
     padding: 0 1rem;
     margin-left: ${({ $sideMenuClosed }) => ($sideMenuClosed ? `-${SIDE_MENU_WIDTH}` : 0)};
     transition: margin 195ms cubic-bezier(0.4, 0, 0.6, 1) 0ms;
+  }
+`;
+
+const SurveyScreenContainer = styled.div<{
+  $scrollable?: boolean;
+}>`
+  display: flex;
+  overflow: ${({ $scrollable }) => ($scrollable ? 'auto' : 'hidden')};
+  align-items: flex-start;
+  height: calc(100vh - ${HEADER_HEIGHT} - ${SURVEY_TOOLBAR_HEIGHT});
+  width: 100vw;
+  ${({ theme }) => theme.breakpoints.up('md')} {
+    margin-left: -1.25rem;
+    padding-top: ${({ $scrollable }) => ($scrollable ? '0' : '2rem')};
+    padding-bottom: 2rem;
   }
 `;
 
@@ -60,43 +72,6 @@ const Form = styled.form`
   position: relative;
 `;
 
-const FormActions = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 0.5rem;
-  border-top: 1px solid ${props => props.theme.palette.divider};
-  button:last-child {
-    margin-left: auto;
-  }
-  ${({ theme }) => theme.breakpoints.up('md')} {
-    padding: 1rem;
-  }
-`;
-
-const ButtonGroup = styled.div`
-  display: flex;
-  button,
-  a {
-    &:last-child {
-      margin-left: 1rem;
-    }
-  }
-`;
-
-const BackButton = styled(Button).attrs({
-  startIcon: <ArrowBackIosIcon />,
-  variant: 'text',
-  color: 'default',
-})`
-  ${({ theme }) => theme.breakpoints.down('md')} {
-    padding-left: 0.8rem;
-    .MuiButton-startIcon {
-      margin-right: 0.25rem;
-    }
-  }
-`;
-
 const LoadingContainer = styled.div`
   width: 100%;
   height: 100%;
@@ -119,43 +94,12 @@ export const SurveyLayout = () => {
     screenNumber,
     sideMenuOpen,
     numberOfScreens,
+    isSuccessScreen,
     isReviewScreen,
-    isResponseScreen,
     visibleScreens,
-    openCancelConfirmation,
   } = useSurveyForm();
   const { handleSubmit, getValues } = useFormContext();
-  const isMobile = useIsMobile();
   const { mutate: submitSurvey, isLoading: isSubmittingSurvey } = useSubmitSurvey();
-
-  const handleStep = (path, data) => {
-    setFormData({ ...formData, ...data });
-    navigate(path);
-  };
-
-  const onStepPrevious = () => {
-    const data = getValues();
-    let path = ROUTES.SURVEY_SELECT;
-    const prevScreenNumber = isReviewScreen ? numberOfScreens : screenNumber! - 1;
-    if (prevScreenNumber) {
-      path = generatePath(ROUTES.SURVEY_SCREEN, {
-        ...params,
-        screenNumber: String(prevScreenNumber),
-      });
-    }
-
-    handleStep(path, data);
-  };
-
-  const navigateNext = data => {
-    const path = isLast
-      ? generatePath(ROUTES.SURVEY_REVIEW, params)
-      : generatePath(ROUTES.SURVEY_SCREEN, {
-          ...params,
-          screenNumber: String(screenNumber! + 1),
-        });
-    handleStep(path, data);
-  };
 
   const onError = errors => {
     // If we're not on the last screen, we don't need to do anything, because the errors get focussed and handled by react-hook-form
@@ -183,6 +127,35 @@ export const SurveyLayout = () => {
     );
   };
 
+  const handleStep = (path, data) => {
+    setFormData({ ...formData, ...data });
+    navigate(path);
+  };
+
+  const navigateNext = data => {
+    const path = isLast
+      ? generatePath(ROUTES.SURVEY_REVIEW, params)
+      : generatePath(ROUTES.SURVEY_SCREEN, {
+          ...params,
+          screenNumber: String(screenNumber! + 1),
+        });
+    handleStep(path, data);
+  };
+
+  const onStepPrevious = () => {
+    const data = getValues();
+    let path = ROUTES.SURVEY_SELECT;
+    const prevScreenNumber = isReviewScreen ? numberOfScreens : screenNumber! - 1;
+    if (prevScreenNumber) {
+      path = generatePath(ROUTES.SURVEY_SCREEN, {
+        ...params,
+        screenNumber: String(prevScreenNumber),
+      });
+    }
+
+    handleStep(path, data);
+  };
+
   const onSubmit = data => {
     if (isReviewScreen) return submitSurvey(data);
     return navigateNext(data);
@@ -190,50 +163,25 @@ export const SurveyLayout = () => {
 
   const handleClickSubmit = handleSubmit(onSubmit, onError);
 
-  const getNextButtonText = () => {
-    if (isReviewScreen) return 'Submit';
-    if (isLast) {
-      return isMobile ? 'Review' : 'Review and submit';
-    }
-    return 'Next';
-  };
-
-  const nextButtonText = getNextButtonText();
-
   return (
     <>
-      <SurveySideMenu />
-      <ScrollableLayout $sideMenuClosed={!sideMenuOpen && !isReviewScreen && !isResponseScreen}>
-        <Paper>
-          <Form onSubmit={handleClickSubmit} noValidate>
-            <Outlet />
-            {isSubmittingSurvey && (
-              <LoadingContainer>
-                <SpinningLoader />
-              </LoadingContainer>
-            )}
-            {!isResponseScreen && (
-              <FormActions>
-                <BackButton onClick={onStepPrevious} disabled={isSubmittingSurvey}>
-                  Back
-                </BackButton>
-                <ButtonGroup>
-                  <Button
-                    onClick={openCancelConfirmation}
-                    variant="outlined"
-                    disabled={isSubmittingSurvey}
-                  >
-                    Cancel
-                  </Button>
-                  <Button type="submit" disabled={isSubmittingSurvey}>
-                    {nextButtonText}
-                  </Button>
-                </ButtonGroup>
-              </FormActions>
-            )}
-          </Form>
-        </Paper>
-      </ScrollableLayout>
+      <SurveyToolbar />
+      <SurveyScreenContainer $scrollable={isSuccessScreen}>
+        <SurveySideMenu />
+        <ScrollableLayout $sideMenuClosed={!sideMenuOpen && !isReviewScreen}>
+          <Paper>
+            <Form onSubmit={handleClickSubmit} noValidate>
+              <Outlet />
+              {isSubmittingSurvey && (
+                <LoadingContainer>
+                  <SpinningLoader />
+                </LoadingContainer>
+              )}
+              <SurveyPaginator onStepPrevious={onStepPrevious} isLoading={isSubmittingSurvey} />
+            </Form>
+          </Paper>
+        </ScrollableLayout>
+      </SurveyScreenContainer>
     </>
   );
 };
