+++ packages/datatrak-web/src/views/SurveyPage.tsx
@@ -9,9 +9,8 @@ import { useForm, FormProvider } from 'react-hook-form';
 import { FullPageLoader } from '@tupaia/ui-components';
 import { useSurvey } from '../api/queries';
 import { CancelConfirmModal } from '../components';
-import { SurveyToolbar, useSurveyForm, useValidationResolver, SurveyContext } from '../features';
+import { useSurveyForm, useValidationResolver, SurveyContext } from '../features';
 import { SurveyParams } from '../types';
-import { HEADER_HEIGHT, SURVEY_TOOLBAR_HEIGHT } from '../constants';
 
 // wrap the entire page so that other content can be centered etc
 const PageWrapper = styled.div`
@@ -23,25 +22,10 @@ const PageWrapper = styled.div`
   }
 `;
 
-const SurveyScreenContainer = styled.div<{
-  $scrollable?: boolean;
-}>`
-  display: flex;
-  overflow: ${({ $scrollable }) => ($scrollable ? 'auto' : 'hidden')};
-  align-items: flex-start;
-  height: calc(100vh - ${HEADER_HEIGHT} - ${SURVEY_TOOLBAR_HEIGHT});
-  width: 100vw;
-  ${({ theme }) => theme.breakpoints.up('md')} {
-    margin-left: -1.25rem;
-    padding-top: ${({ $scrollable }) => ($scrollable ? '0' : '2rem')};
-    padding-bottom: 2rem;
-  }
-`;
-
 const SurveyPageInner = () => {
   const { surveyCode, screenNumber } = useParams<SurveyParams>();
   const { isLoading } = useSurvey(surveyCode);
-  const { formData, isSuccessScreen, cancelModalOpen, closeCancelConfirmation } = useSurveyForm();
+  const { formData, cancelModalOpen, closeCancelConfirmation } = useSurveyForm();
   const resolver = useValidationResolver();
   const formContext = useForm({ defaultValues: formData, reValidateMode: 'onSubmit', resolver });
 
@@ -52,12 +36,9 @@ const SurveyPageInner = () => {
   return (
     <PageWrapper>
       <FormProvider {...formContext}>
-        <SurveyToolbar />
-        <SurveyScreenContainer $scrollable={isSuccessScreen}>
-          {/* Use a key to render a different survey screen component for every screen number. This is so
+        {/* Use a key to render a different survey screen component for every screen number. This is so
       that the screen can be easily initialised with the form data. See https://react.dev/learn/you-might-not-need-an-effect#resetting-all-state-when-a-prop-changes */}
-          <Outlet key={screenNumber} />
-        </SurveyScreenContainer>
+        <Outlet key={screenNumber} />
       </FormProvider>
       <CancelConfirmModal isOpen={cancelModalOpen} onClose={closeCancelConfirmation} />
     </PageWrapper>
