+++ packages/types/src/types/requests/datatrak-web-server/SingleSurveyResponseRequest.ts
@@ -3,7 +3,7 @@
  * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
-import { SurveyResponse } from '../../models';
+import { Country, Entity, Survey, SurveyResponse } from '../../models';
 import { KeysToCamelCase } from '../../../utils/casing';
 
 export type Params = {
@@ -12,6 +12,10 @@ export type Params = {
 
 export interface ResBody extends KeysToCamelCase<SurveyResponse> {
   answers: Record<string, string>;
+  countryName: Country['name'];
+  entityName: Entity['name'];
+  surveyName: Survey['name'];
+  surveyCode: Survey['code'];
 }
 export type ReqBody = Record<string, never>;
 export type ReqQuery = Record<string, never>;
